Rename and update OpenMX test to reflect multiple format support
